+++ rustuna_pyo3/src/lib.rs
@@ -36,7 +36,8 @@ fn rustuna(_py: Python<'_>, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<storage::PyStorage>()?;
     m.add_class::<pyobject_storage::PyPyObjectStorage>()?;
     // sampler
-    m.add_class::<sampler::PySampler>()?;
+    m.add_class::<sampler::tpe::PyTpeSampler>()?;
+    m.add_class::<sampler::nsgaii::PyNSGAIISampler>()?;
     m.add_class::<sampler::PySamplerContext>()?;
     m.add_class::<sampler::cmaes::PyCmaEsSampler>()?;
     m.add_class::<sampler::random::PyRandomSampler>()?;
