+++ rustuna_pyo3/src/sampler/tpe.rs
@@ -1,0 +1,111 @@
+use std::collections::HashMap;
+use std::sync::{Arc, Mutex};
+
+use pyo3::exceptions::PyRuntimeError;
+use pyo3::prelude::*;
+use pyo3::Py;
+
+use rustuna_core::sampler::Sampler;
+use rustuna_core::trial::TrialStateValues;
+use rustuna_sampler::tpe::{TpeConfig, TpeSampler};
+
+use crate::distribution::PyDistribution;
+use crate::sampler::{extract_storage, PySamplerContext};
+use crate::trial::PyTrialState;
+
+#[derive(Clone)]
+#[pyclass(name = "TPESampler")]
+#[pyo3(module = "rustuna")]
+pub struct PyTpeSampler {
+    pub sampler: Arc<Mutex<TpeSampler>>,
+}
+#[pymethods]
+impl PyTpeSampler {
+    #[new]
+    #[pyo3(signature = (*, seed = None, n_startup_trials = 10, multivariate = true))]
+    fn py_new(seed: Option<u64>, n_startup_trials: usize, multivariate: bool) -> PyResult<Self> {
+        let rs_sampler = TpeSampler::from_config(TpeConfig {
+            seed,
+            n_startup_trials,
+            multivariate,
+        });
+        Ok(PyTpeSampler {
+            sampler: Arc::new(Mutex::new(rs_sampler)),
+        })
+    }
+
+    #[getter]
+    fn support_joint_sampling(&self) -> PyResult<bool> {
+        let guard = self
+            .sampler
+            .lock()
+            .map_err(|_| PyRuntimeError::new_err("Failed to acquire sampler lock"))?;
+        Ok(guard.support_joint_sampling())
+    }
+
+    fn sample_independent(
+        &self,
+        ctx: &PySamplerContext,
+        storage: Py<PyAny>,
+        name: &str,
+        distribution: &PyDistribution,
+    ) -> PyResult<f64> {
+        let arc_storage = extract_storage(storage)?;
+        self.sampler
+            .lock()
+            .map_err(|_| PyRuntimeError::new_err("Failed to acquire the sampler guard"))?
+            .sample_independent(
+                &ctx.context.clone(),
+                arc_storage,
+                name,
+                &distribution.distribution,
+            )
+            .map_err(|e| PyRuntimeError::new_err(format!("Failed to sample independent: {e:?}")))
+    }
+
+    fn sample_joint(
+        &self,
+        ctx: &PySamplerContext,
+        storage: Py<PyAny>,
+        search_space: HashMap<String, PyDistribution>,
+    ) -> PyResult<HashMap<String, f64>> {
+        let arc_storage = extract_storage(storage)?;
+        self.sampler
+            .lock()
+            .map_err(|_| PyRuntimeError::new_err("Failed to acquire the sampler guard"))?
+            .sample_joint(
+                &ctx.context.clone(),
+                arc_storage,
+                &search_space
+                    .into_iter()
+                    .map(|(k, v)| (k, v.distribution))
+                    .collect(),
+            )
+            .map_err(|e| PyRuntimeError::new_err(format!("Failed to sample joint: {e:?}")))
+    }
+
+    #[pyo3(signature = (ctx, storage, state, values = None))]
+    fn after_trial(
+        &self,
+        ctx: &PySamplerContext,
+        storage: Py<PyAny>,
+        state: PyTrialState,
+        values: Option<Vec<f64>>,
+    ) -> PyResult<()> {
+        let arc_storage = extract_storage(storage)?;
+        let state_values = match state {
+            PyTrialState::RUNNING => TrialStateValues::Running,
+            PyTrialState::COMPLETE => TrialStateValues::Complete(values.ok_or(
+                PyRuntimeError::new_err("values must be specified when state is COMPLETE"),
+            )?),
+            PyTrialState::PRUNED => TrialStateValues::Pruned,
+            PyTrialState::WAITING => TrialStateValues::Waiting,
+            PyTrialState::FAIL => TrialStateValues::Fail,
+        };
+        self.sampler
+            .lock()
+            .map_err(|_| PyRuntimeError::new_err("Failed to acquire the sampler guard"))?
+            .after_trial(&ctx.context, arc_storage, &state_values)
+            .map_err(|e| PyRuntimeError::new_err(format!("Failed to call after_trial: {e:?}")))
+    }
+}
